feature: diseño login, añadido manejo de modulos en main
