+++ src/app/protected/profile/page.tsx
@@ -1,0 +1,10 @@
+export default function PerfilPage() {
+  return (
+    <div className="p-6">
+      <h1 className="text-2xl font-bold text-gray-800 mb-4">Perfil</h1>
+      <p className="text-gray-600">
+        Aquí irá la edición del perfil del usuario.
+      </p>
+    </div>
+  );
+}
